+++ example/json/exampleMessages_cs.i69n.dart
@@ -3,22 +3,34 @@
 // dart format off
 import 'package:i69n/i69n.dart' as i69n;
 import 'exampleMessages.i69n.dart';
+import '../gender.dart';
 
 String get _languageCode => 'cs';
 String get _localeName => 'cs';
 
-String _plural(int count, {String? zero, String? one, String? two, String? few, String? many, String? other}) =>
-    i69n.plural(count, _languageCode, zero: zero, one: one, two: two, few: few, many: many, other: other);
+String _plural(int count,
+        {String? zero,
+        String? one,
+        String? two,
+        String? few,
+        String? many,
+        String? other}) =>
+    i69n.plural(count, _languageCode,
+        zero: zero, one: one, two: two, few: few, many: many, other: other);
+String _select(Object? value, Map<String, String> cases) =>
+    i69n.select(value, cases);
 
 class ExampleMessages_cs extends ExampleMessages {
   const ExampleMessages_cs();
   GenericExampleMessages_cs get generic => GenericExampleMessages_cs(this);
   InvoiceExampleMessages_cs get invoice => InvoiceExampleMessages_cs(this);
   ApplesExampleMessages_cs get apples => ApplesExampleMessages_cs(this);
+  PersonExampleMessages_cs get person => PersonExampleMessages_cs(this);
   Object operator [](String key) {
     var index = key.indexOf('.');
     if (index > 0) {
-      return (this[key.substring(0, index)] as i69n.I69nMessageBundle)[key.substring(index + 1)];
+      return (this[key.substring(0, index)]
+          as i69n.I69nMessageBundle)[key.substring(index + 1)];
     }
     switch (key) {
       case 'generic':
@@ -27,6 +39,8 @@ class ExampleMessages_cs extends ExampleMessages {
         return invoice;
       case 'apples':
         return apples;
+      case 'person':
+        return person;
       default:
         return super[key];
     }
@@ -40,7 +54,8 @@ class GenericExampleMessages_cs extends GenericExampleMessages {
   Object operator [](String key) {
     var index = key.indexOf('.');
     if (index > 0) {
-      return (this[key.substring(0, index)] as i69n.I69nMessageBundle)[key.substring(index + 1)];
+      return (this[key.substring(0, index)]
+          as i69n.I69nMessageBundle)[key.substring(index + 1)];
     }
     switch (key) {
       case 'done':
@@ -57,11 +72,13 @@ class InvoiceExampleMessages_cs extends InvoiceExampleMessages {
   String get create => "Vytvořit fakturu";
   String get delete => "Smazat fakturu";
   String get help => "Tuhle funkci použij na vytváření faktur. Boží!";
-  String count(int a) => "Už jsi vytvořil ${_plural(a, one: 'fakturu', few: 'faktury', many: 'faktur')}.";
+  String count(int a) =>
+      "Už jsi vytvořil ${_plural(a, one: 'fakturu', few: 'faktury', many: 'faktur')}.";
   Object operator [](String key) {
     var index = key.indexOf('.');
     if (index > 0) {
-      return (this[key.substring(0, index)] as i69n.I69nMessageBundle)[key.substring(index + 1)];
+      return (this[key.substring(0, index)]
+          as i69n.I69nMessageBundle)[key.substring(index + 1)];
     }
     switch (key) {
       case 'create':
@@ -87,13 +104,44 @@ class ApplesExampleMessages_cs extends ApplesExampleMessages {
   Object operator [](String key) {
     var index = key.indexOf('.');
     if (index > 0) {
-      return (this[key.substring(0, index)] as i69n.I69nMessageBundle)[key.substring(index + 1)];
+      return (this[key.substring(0, index)]
+          as i69n.I69nMessageBundle)[key.substring(index + 1)];
     }
     switch (key) {
       case '_apples':
         return _apples;
       case 'count':
         return count;
+      default:
+        return super[key];
+    }
+  }
+}
+
+class PersonExampleMessages_cs extends PersonExampleMessages {
+  final ExampleMessages_cs _parent;
+  const PersonExampleMessages_cs(this._parent) : super(_parent);
+  String sees(Gender gender) => "Vidím ${_select(gender, {
+            'male': 'ho',
+            'female': 'ji',
+            'other': 'je'
+          })}.";
+  String owns(Gender gender, int cnt) => "${_select(gender, {
+            'male': 'Jeho',
+            'female': 'Její',
+            'other': 'Jejich'
+          })} ${_plural(cnt, one: 'jablko', few: 'jablka', many: 'jablek')}.";
+  Object operator [](String key) {
+    var index = key.indexOf('.');
+    if (index > 0) {
+      return (this[key.substring(0, index)]
+          as i69n.I69nMessageBundle)[key.substring(index + 1)];
+    }
+    switch (key) {
+      case 'sees':
+        return sees;
+      case 'owns':
+        return owns;
       default:
         return super[key];
     }