+++ test/mock/remoteMessages.i69n.dart
@@ -2,6 +2,7 @@
 // GENERATED FILE, do not edit!
 // dart format off
 import 'package:i69n/i69n.dart' as i69n;
+import 'gender.dart';
 
 String get _languageCode => 'en';
 String get _localeName => 'en';
@@ -10,6 +11,8 @@ const Map<String, String> _baked = {
   'title': "Welcome",
   'greeting': "Hi \$name",
   'apples': "\${_plural(count, one: '\$count apple', other: '\$count apples')}",
+  'sees':
+      "I see \${_select(gender, male: 'Him', female: 'Her', other: 'Them')}",
   'home.subtitle': "Home",
 };
 
@@ -23,14 +26,20 @@ class RemoteMessages implements i69n.I69nMessageBundle {
   }
 
   Map<String, String> get i69nRemoteMessages => i69nRemoteData;
-  String get title => i69n.tr(i69nRemoteMessages, _baked, 'title', const {}, _languageCode);
-  String greeting(String name) => i69n.tr(i69nRemoteMessages, _baked, 'greeting', {'name': name}, _languageCode);
-  String apples(int count) => i69n.tr(i69nRemoteMessages, _baked, 'apples', {'count': count}, _languageCode);
+  String get title =>
+      i69n.tr(i69nRemoteMessages, _baked, 'title', const {}, _languageCode);
+  String greeting(String name) => i69n.tr(
+      i69nRemoteMessages, _baked, 'greeting', {'name': name}, _languageCode);
+  String apples(int count) => i69n.tr(
+      i69nRemoteMessages, _baked, 'apples', {'count': count}, _languageCode);
+  String sees(Gender gender) => i69n.tr(
+      i69nRemoteMessages, _baked, 'sees', {'gender': gender}, _languageCode);
   HomeRemoteMessages get home => HomeRemoteMessages(this);
   Object operator [](String key) {
     var index = key.indexOf('.');
     if (index > 0) {
-      return (this[key.substring(0, index)] as i69n.I69nMessageBundle)[key.substring(index + 1)];
+      return (this[key.substring(0, index)]
+          as i69n.I69nMessageBundle)[key.substring(index + 1)];
     }
     switch (key) {
       case 'title':
@@ -39,6 +48,8 @@ class RemoteMessages implements i69n.I69nMessageBundle {
         return greeting;
       case 'apples':
         return apples;
+      case 'sees':
+        return sees;
       case 'home':
         return home;
       default:
@@ -51,11 +62,13 @@ class HomeRemoteMessages implements i69n.I69nMessageBundle {
   final RemoteMessages _parent;
   const HomeRemoteMessages(this._parent);
   Map<String, String> get i69nRemoteMessages => _parent.i69nRemoteMessages;
-  String get subtitle => i69n.tr(i69nRemoteMessages, _baked, 'home.subtitle', const {}, _languageCode);
+  String get subtitle => i69n.tr(
+      i69nRemoteMessages, _baked, 'home.subtitle', const {}, _languageCode);
   Object operator [](String key) {
     var index = key.indexOf('.');
     if (index > 0) {
-      return (this[key.substring(0, index)] as i69n.I69nMessageBundle)[key.substring(index + 1)];
+      return (this[key.substring(0, index)]
+          as i69n.I69nMessageBundle)[key.substring(index + 1)];
     }
     switch (key) {
       case 'subtitle':