+++ test/testMessages.i69n.dart
@@ -2,28 +2,44 @@
 // GENERATED FILE, do not edit!
 // dart format off
 import 'package:i69n/i69n.dart' as i69n;
-import 'dart:io';
+import 'mock/gender.dart';
 
 String get _languageCode => 'sk';
 String get _localeName => 'en';
 
-String _plural(int count, {String? zero, String? one, String? two, String? few, String? many, String? other}) =>
-    i69n.plural(count, _languageCode, zero: zero, one: one, two: two, few: few, many: many, other: other);
-String _ordinal(int count, {String? zero, String? one, String? two, String? few, String? many, String? other}) =>
-    i69n.ordinal(count, _languageCode, zero: zero, one: one, two: two, few: few, many: many, other: other);
-String _cardinal(int count, {String? zero, String? one, String? two, String? few, String? many, String? other}) =>
-    i69n.cardinal(count, _languageCode, zero: zero, one: one, two: two, few: few, many: many, other: other);
+String _plural(int count,
+        {String? zero,
+        String? one,
+        String? two,
+        String? few,
+        String? many,
+        String? other}) =>
+    i69n.plural(count, _languageCode,
+        zero: zero, one: one, two: two, few: few, many: many, other: other);
+String _ordinal(int count,
+        {String? zero,
+        String? one,
+        String? two,
+        String? few,
+        String? many,
+        String? other}) =>
+    i69n.ordinal(count, _languageCode,
+        zero: zero, one: one, two: two, few: few, many: many, other: other);
+String _select(Object? value, Map<String, String> cases) =>
+    i69n.select(value, cases);
 
 class TestMessages implements i69n.I69nMessageBundle {
   const TestMessages();
   GenericTestMessages get generic => GenericTestMessages(this);
   InvoiceTestMessages get invoice => InvoiceTestMessages(this);
   ApplesTestMessages get apples => ApplesTestMessages(this);
+  PersonTestMessages get person => PersonTestMessages(this);
   FriendsTestMessages get friends => FriendsTestMessages(this);
   Object operator [](String key) {
     var index = key.indexOf('.');
     if (index > 0) {
-      return (this[key.substring(0, index)] as i69n.I69nMessageBundle)[key.substring(index + 1)];
+      return (this[key.substring(0, index)]
+          as i69n.I69nMessageBundle)[key.substring(index + 1)];
     }
     switch (key) {
       case 'generic':
@@ -32,6 +48,8 @@ class TestMessages implements i69n.I69nMessageBundle {
         return invoice;
       case 'apples':
         return apples;
+      case 'person':
+        return person;
       case 'friends':
         return friends;
       default:
@@ -46,11 +64,13 @@ class GenericTestMessages implements i69n.I69nMessageBundle {
   String get ok => "OK";
   String get done => "DONE";
   String get letsGo => "Let's go!";
-  String ordinalNumber(int n) => "${_ordinal(n, one: '1st', two: '2nd', few: '3rd', other: '${n}th')}";
+  String ordinalNumber(int n) =>
+      "${_ordinal(n, one: '1st', two: '2nd', few: '3rd', other: '${n}th')}";
   Object operator [](String key) {
     var index = key.indexOf('.');
     if (index > 0) {
-      return (this[key.substring(0, index)] as i69n.I69nMessageBundle)[key.substring(index + 1)];
+      return (this[key.substring(0, index)]
+          as i69n.I69nMessageBundle)[key.substring(index + 1)];
     }
     switch (key) {
       case 'ok':
@@ -72,22 +92,27 @@ class InvoiceTestMessages implements i69n.I69nMessageBundle {
   const InvoiceTestMessages(this._parent);
   String get create => "Create invoice";
   String get delete => "Delete  invoice";
-  String get help => "Use this function to generate new invoices and stuff. Awesome!";
-  String count(int cnt) => "You have created $cnt ${_plural(cnt, one: 'invoice', many: 'invoices')}.";
+  String get help =>
+      "Use this function to generate new invoices and stuff. Awesome!";
+  String count(int cnt) =>
+      "You have created $cnt ${_plural(cnt, one: 'invoice', many: 'invoices')}.";
   String get something => "Let\'s go!";
   Object operator [](String key) {
     var index = key.indexOf('.');
     if (index > 0) {
-      return (this[key.substring(0, index)] as i69n.I69nMessageBundle)[key.substring(index + 1)];
+      return (this[key.substring(0, index)]
+          as i69n.I69nMessageBundle)[key.substring(index + 1)];
     }
-    throw Exception('[] operator is disabled in en.invoice, see _i69n: nomap flag.');
+    throw Exception(
+        '[] operator is disabled in sk.invoice, see _i69n: nomap flag.');
   }
 }
 
 class ApplesTestMessages implements i69n.I69nMessageBundle {
   final TestMessages _parent;
   const ApplesTestMessages(this._parent);
-  String _apples(int cnt) => "${_plural(cnt, zero: 'no apples', one: '$cnt apple', many: '$cnt apples')}";
+  String _apples(int cnt) =>
+      "${_plural(cnt, zero: 'no apples', one: '$cnt apple', many: '$cnt apples')}";
   String count(int cnt) => "You have eaten ${_apples(cnt)}.";
   String problematic(int count) =>
       "${_plural(count, zero: 'didn\'t find any tasks', one: 'found 1 task', other: 'found $count tasks')}";
@@ -97,7 +122,8 @@ class ApplesTestMessages implements i69n.I69nMessageBundle {
   Object operator [](String key) {
     var index = key.indexOf('.');
     if (index > 0) {
-      return (this[key.substring(0, index)] as i69n.I69nMessageBundle)[key.substring(index + 1)];
+      return (this[key.substring(0, index)]
+          as i69n.I69nMessageBundle)[key.substring(index + 1)];
     }
     switch (key) {
       case '_apples':
@@ -118,6 +144,36 @@ class ApplesTestMessages implements i69n.I69nMessageBundle {
   }
 }
 
+class PersonTestMessages implements i69n.I69nMessageBundle {
+  final TestMessages _parent;
+  const PersonTestMessages(this._parent);
+  String sees(Gender gender) => "I see ${_select(gender, {
+            'male': 'Him',
+            'female': 'Her',
+            'other': 'Them'
+          })}";
+  String owns(Gender gender, int cnt) => "${_select(gender, {
+            'male': 'his',
+            'female': 'her',
+            'other': 'their'
+          })} ${_plural(cnt, one: '$cnt apple', many: '$cnt apples')}";
+  Object operator [](String key) {
+    var index = key.indexOf('.');
+    if (index > 0) {
+      return (this[key.substring(0, index)]
+          as i69n.I69nMessageBundle)[key.substring(index + 1)];
+    }
+    switch (key) {
+      case 'sees':
+        return sees;
+      case 'owns':
+        return owns;
+      default:
+        return key;
+    }
+  }
+}
+
 class FriendsTestMessages implements i69n.I69nMessageBundle {
   final TestMessages _parent;
   const FriendsTestMessages(this._parent);
@@ -126,7 +182,8 @@ class FriendsTestMessages implements i69n.I69nMessageBundle {
   Object operator [](String key) {
     var index = key.indexOf('.');
     if (index > 0) {
-      return (this[key.substring(0, index)] as i69n.I69nMessageBundle)[key.substring(index + 1)];
+      return (this[key.substring(0, index)]
+          as i69n.I69nMessageBundle)[key.substring(index + 1)];
     }
     switch (key) {
       case 'michael':
@@ -147,7 +204,8 @@ class MichaelFriendsTestMessages implements i69n.I69nMessageBundle {
   Object operator [](String key) {
     var index = key.indexOf('.');
     if (index > 0) {
-      return (this[key.substring(0, index)] as i69n.I69nMessageBundle)[key.substring(index + 1)];
+      return (this[key.substring(0, index)]
+          as i69n.I69nMessageBundle)[key.substring(index + 1)];
     }
     switch (key) {
       case 'name':
@@ -160,20 +218,19 @@ class MichaelFriendsTestMessages implements i69n.I69nMessageBundle {
   }
 }
 
-class EvaFriendsTestMessages implements i69n.I69nMessageBundle, MichaelFriendsTestMessages {
+class EvaFriendsTestMessages
+    implements i69n.I69nMessageBundle, MichaelFriendsTestMessages {
   final FriendsTestMessages _parent;
   const EvaFriendsTestMessages(this._parent);
-  String get _i69n_implements => "MichaelFriendsTestMessages";
   String get name => "Bbbbb";
   String get description => "Bb Bb Bb";
   Object operator [](String key) {
     var index = key.indexOf('.');
     if (index > 0) {
-      return (this[key.substring(0, index)] as i69n.I69nMessageBundle)[key.substring(index + 1)];
+      return (this[key.substring(0, index)]
+          as i69n.I69nMessageBundle)[key.substring(index + 1)];
     }
     switch (key) {
-      case '_i69n_implements':
-        return _i69n_implements;
       case 'name':
         return name;
       case 'description':